+++ HISTORY.markdown
@@ -2,6 +2,7 @@
 
 ## unreleased
 * Added propagation of errors from data source
+* API improvement: No need to call <tt>results</tt> on FoodItem.search to access search results
 
 ## 0.0.1  September 10, 2011
 * Initial release. FatSecret adapter supports both <tt>search</tt> and <tt>details</tt> methods.